requirements.txt
mysql-replication
psycopg2
